added landing page
