+++ src/app/globals.css
@@ -1,8 +1,11 @@
-@import 'tailwindcss';
 @import url('https://fonts.googleapis.com/css2?family=Geist:wght@100..900&display=swap');
+@import 'tailwindcss';
 
+html,
 body {
-  font-family: 'Geist', sans-serif;
+  height: 100%;
+  margin: 0;
+  padding: 0;
 }
 
 @theme {
